migrate hash methods to new hash.go
